Moves clean_text() to clean module.
