+++ eyecite/__init__.py
@@ -1,7 +1,8 @@
 from .annotate import annotate_citations
+from .clean import clean_text
 from .find_citations import get_citations
 from .resolve import resolve_citations
-from .utils import clean_text, dump_citations
+from .utils import dump_citations
 
 __all__ = [
     "annotate_citations",
