+++ eyecite/utils.py
@@ -1,9 +1,6 @@
 import re
-from typing import Callable, Iterable, Union
 
 from lxml import etree
-
-from eyecite.cleaners import cleaners_lookup
 
 
 def strip_punct(text: str) -> str:
@@ -35,25 +32,6 @@ def strip_punct(text: str) -> str:
     text = re.sub(r"(\S)(\'\'?)", r"\1", text)
 
     return text.strip()
-
-
-def clean_text(text, steps: Iterable[Union[str, Callable[[str], str]]]) -> str:
-    """Applies each step in order to text, returning the result.
-    Steps may be the names of functions in eyecite.cleaners, or callables.
-    """
-    for step in steps:
-        if step in cleaners_lookup:
-            step_func = cleaners_lookup[step]  # type: ignore
-        elif callable(step):
-            step_func = step
-        else:
-            raise ValueError(
-                "clean_text steps must be callable "
-                f"or one of {list(cleaners_lookup.keys())}"
-            )
-        text = step_func(text)
-
-    return text  # type: ignore
 
 
 def is_balanced_html(text: str) -> bool:
